% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{swi_pal}
\alias{swi_col}
\alias{swi_dpal}
\alias{swi_dpal2}
\alias{swi_dpal3}
\alias{swi_pal}
\alias{swi_rpal}
\alias{swi_spal}
\title{swissinfo.ch standard color palette}
\format{An object of class \code{character} of length 22.}
\usage{
swi_pal

swi_col

swi_rpal

swi_dpal

swi_dpal2

swi_dpal3

swi_spal
}
\description{
Just a bunch of hex color vectors... 
Color palette as vector (no function). Colors can be sequential (one color), diverging (2 colors)

swissinfo 2017 divergent color palette!

swissinfo random color palette
}
\details{
swi_pal: all SWI colors (including sequential beige)

swi_col: SWI 2017 divergent color palette! 6 levels

swi_rpal: all SWI colors, same as swi_pal but randomized

swi_dpal: diverging color palette with 10 levels based on interactive things color scheme for the federal ballot

swi_dpal2: another diverging color palette, 11 levels. Same as swi_dpal , but with white (very light grey) in the middle

swi_dpal3: another diverging color palette, 19 levels. Same as swi_dpal2

swi_spal:  single hue beige color palette, 6 factors
}
\examples{
pie(rep(1,length(swi_pal)), col=swi_pal)
pie(rep(1,length(swi_col)), col=swi_col)
pie(rep(1,length(swi_rpal)), col=swi_rpal)
pie(rep(1,length(swi_dpal)), col=swi_dpal)
pie(rep(1,length(swi_dpal2)), col=swi_dpal2)
pie(rep(1,length(swi_dpal3)), col=swi_dpal3)
pie(rep(1,length(swi_spal)), col=swi_spal)
}
\keyword{datasets}

