% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_helper.R
\name{loadTranslation}
\alias{loadTranslation}
\title{Simple convenience function to load text data}
\usage{
loadTranslation(path, ...)
}
\arguments{
\item{path, }{a string path to a csv file with}
}
\description{
loadTranslation(): load a csv file and discard verbosely incomplete columns
}

