% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}

\item{\dots}{ggplot objects}
}
\description{
To arrange multiple ggplot chart on a grid. Copied from: http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then plot 1 will go in the upper left, 2 will go in the upper right, and ll go all the way across the bottom.
}
\examples{
q1 <- qplot(1:10, 1:10, size = 10:1) + xlab("axis x label") + ylab ("y axis label") + theme_swi2()
q2 <- qplot(mpg, data = mtcars, geom = "dotplot") + theme_swi()
multiplot(q1, q2)
}

