% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_swi.R
\name{pdfswi_sq}
\alias{pdfswi_long}
\alias{pdfswi_sq}
\title{swissinfo standard chart size pdf export}
\usage{
pdfswi_sq(file = "", width = 10, height = 10, useDingbats = FALSE, ...)

pdfswi_long(file = "", width = 6, height = width * 1.25,
  useDingbats = FALSE, ...)
}
\arguments{
\item{file}{a character string giving the name of the file.
    If it is of the form \code{"|cmd"}, the output is piped to the
    command given by \command{cmd}.  If it is \code{NULL},
    then no external file is created (effectively, no drawing occurs),
    but the device may still be queried (e.g., for size of text).

    For use with \code{onefile = FALSE} give a C integer format such
    as \code{"Rplot\%03d.pdf"} (the default in that case).
    (See \code{\link{postscript}} for further details.)

    Tilde expansion (see \code{\link{path.expand}}) is done.
  }

\item{width}{the width and height of the graphics region in
    inches.  The default values are \code{7}.}

\item{useDingbats}{logical.  Should small circles be rendered
    \emph{via} the Dingbats font?  Defaults to \code{TRUE}, which produces
    smaller and better output.  Setting this to \code{FALSE} can work
    around font display problems in broken PDF viewers: although this
    font is one of the 14 guaranteed to be available in all PDF viewers,
    that guarantee is not always honoured.
    See the \sQuote{Note} for a possible fix for some viewers.
    }
}
\description{
Simple wrapper for R's graphics device driver (i.e. pdf) to save charts in pre-defined sizes
}

