% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{swi_pal}
\alias{swi_dpal}
\alias{swi_pal}
\alias{swi_rpal}
\title{swissinfo.ch standard color palette}
\format{\preformatted{ chr [1:22] "#336666" "#368596" "#669999" "#366096" "#333366" ...
}}
\usage{
swi_pal

swi_rpal

swi_dpal
}
\description{
swissinfo.ch standard color palette

swissinfo random color palette
}
\examples{
pie(rep(1,length(swi_pal)), col=swi_pal)
pie(rep(1,length(swi_rpal)), col=swi_rpal)
pie(rep(1,length(swi_dpal)), col=swi_dpal)
}
\keyword{datasets}

