% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_swi.R
\name{pdfswi_sq}
\alias{pdfswi_long}
\alias{pdfswi_sq}
\title{swissinfo standard chart size pdf export}
\usage{
pdfswi_sq(file = "", width = 10, height = 10, useDingbats = FALSE, ...)

pdfswi_long(file = "", width = 6, height = width * 1.25,
  useDingbats = FALSE, ...)
}
\description{
Simple wrapper for R's graphics device driver (i.e. pdf) to save charts in pre-defined sizes
}

