% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pdfswi_sq}
\alias{pdfswi_long}
\alias{pdfswi_sq}
\title{swissinfo standard chart size pdf export}
\usage{
pdfswi_sq(file = "", width = 10, height = 10, ...)

pdfswi_long(file = "", width = 6, height = width * 1.25, ...)
}
\description{
Simple wrapper for R's graphics device driver (i.e. pdf) to save charts in pre-defined sizes
}

